+++ backend/sokdak/src/main/java/com/wooteco/sokdak/member/dto/NicknameUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.wooteco.sokdak.profile.dto;
+package com.wooteco.sokdak.member.dto;
 
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
