+++ backend/sokdak/src/main/java/com/wooteco/sokdak/post/service/PostService.java
@@ -20,7 +20,7 @@ import com.wooteco.sokdak.post.dto.PostUpdateRequest;
 import com.wooteco.sokdak.post.dto.PostsResponse;
 import com.wooteco.sokdak.post.exception.PostNotFoundException;
 import com.wooteco.sokdak.post.repository.PostRepository;
-import com.wooteco.sokdak.profile.dto.MyPostsResponse;
+import com.wooteco.sokdak.post.dto.MyPostsResponse;
 import java.util.Collections;
 import java.util.List;
 import org.springframework.data.domain.Page;
