+++ backend/sokdak/src/test/java/com/wooteco/sokdak/member/service/MemberServiceTest.java
@@ -12,8 +12,8 @@ import com.wooteco.sokdak.member.dto.UniqueResponse;
 import com.wooteco.sokdak.member.exception.InvalidNicknameException;
 import com.wooteco.sokdak.member.repository.AuthCodeRepository;
 import com.wooteco.sokdak.member.repository.MemberRepository;
-import com.wooteco.sokdak.profile.dto.NicknameUpdateRequest;
-import com.wooteco.sokdak.profile.exception.DuplicateNicknameException;
+import com.wooteco.sokdak.member.dto.NicknameUpdateRequest;
+import com.wooteco.sokdak.member.exception.DuplicateNicknameException;
 import com.wooteco.sokdak.util.IntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
