+++ backend/sokdak/src/test/java/com/wooteco/sokdak/post/service/PostServiceTest.java
@@ -26,7 +26,7 @@ import com.wooteco.sokdak.post.dto.PostsElementResponse;
 import com.wooteco.sokdak.post.dto.PostsResponse;
 import com.wooteco.sokdak.post.exception.PostNotFoundException;
 import com.wooteco.sokdak.post.repository.PostRepository;
-import com.wooteco.sokdak.profile.dto.MyPostsResponse;
+import com.wooteco.sokdak.post.dto.MyPostsResponse;
 import com.wooteco.sokdak.util.IntegrationTest;
 import java.util.ArrayList;
 import java.util.Collections;
